requirements.txt
amqp==5.2.0
asgiref==3.7.2
billiard==4.2.0
cachetools==5.3.2
celery==5.3.5
certifi==2023.11.17
charset-normalizer==3.3.2
click==8.1.7
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.3.0
Django==4.2.7
django-cors-headers==4.3.1
djangorestframework==3.14.0
google-api-core==2.14.0
google-api-python-client==2.108.0
google-auth==2.23.4
google-auth-httplib2==0.1.1
googleapis-common-protos==1.61.0
gunicorn==21.2.0
httplib2==0.22.0
idna==3.4
kombu==5.3.4
oauth2client==4.1.3
packaging==23.2
prompt-toolkit==3.0.41
protobuf==4.25.1
psycopg2-binary==2.9.9
pyasn1==0.5.1
pyasn1-modules==0.3.0
PyDrive==1.3.1
pyparsing==3.1.1
python-dateutil==2.8.2
python-dotenv==1.0.0
pytz==2023.3.post1
PyYAML==6.0.1
redis==4.6.0
requests==2.31.0
rsa==4.9
six==1.16.0
sqlparse==0.4.4
tzdata==2023.3
uritemplate==4.1.1
urllib3==2.1.0
vine==5.1.0
wcwidth==0.2.12
